Add: kill() function and auxiliary txt
With the objective of making automatic the multiprocess() function so that it can scrape every url in prensa's jsons, the os.system() command was replaced with subprocess.Popen() command and a kill function was included. The kill functions kills the launched processes that scrape newspapers so that another newspapers can be scraped.

The auxiliary txt is needed to save the proceses pids.